fix: CodeRabbit レビュー指摘の修正（PR #415）
- actor_cell: Kill system経路でFailureMessageSnapshotを生成するよう修正
- cluster_api: AsSnapshot購読でsubscribe→snapshot順序に変更しイベント欠落防止
- cluster_core: start_member/start_clientでlast_topology_hashをリセット
- cluster_core/tests: FailingProviderのjoin_fail/leave_fail未使用コード削除
- cluster_event_type: ワイルドカード意図のコメント追加
- local_cluster_provider: 重複joinでearly returnを追加
- cluster_router_pool: total_instancesをnext_routeeに反映
- remote config: ack_send_windowをusize→u64に統一（全使用箇所）

via [Happy](https://happy.engineering)

+++ modules/remote/src/core/remoting_extension/config.rs
@@ -28,7 +28,7 @@ pub struct RemotingExtensionConfig {
   auto_start:               bool,
   handshake_timeout:        Duration,
   shutdown_flush_timeout:   Duration,
-  ack_send_window:          usize,
+  ack_send_window:          u64,
   ack_receive_window:       u64,
   transport_scheme:         String,
   backpressure_listeners:   Vec<Box<dyn RemotingBackpressureListener>>,
@@ -128,7 +128,7 @@ impl RemotingExtensionConfig {
   ///
   /// Panics when `window` is zero.
   #[must_use]
-  pub fn with_ack_send_window(mut self, window: usize) -> Self {
+  pub fn with_ack_send_window(mut self, window: u64) -> Self {
     assert!(window > 0, "ack send window must be > 0");
     self.ack_send_window = window;
     self
@@ -209,7 +209,7 @@ impl RemotingExtensionConfig {
 
   /// Returns the configured outbound ack send window.
   #[must_use]
-  pub const fn ack_send_window(&self) -> usize {
+  pub const fn ack_send_window(&self) -> u64 {
     self.ack_send_window
   }
