+++ modules/remote/src/std/endpoint_transport_bridge/config.rs
@@ -37,7 +37,7 @@ pub struct EndpointTransportBridgeConfig<TB: RuntimeToolbox + 'static> {
   /// Timeout applied when flushing outstanding messages during graceful shutdown.
   pub shutdown_flush_timeout: Duration,
   /// Maximum number of outbound system messages kept pending per authority.
-  pub ack_send_window:        usize,
+  pub ack_send_window:        u64,
   /// Maximum accepted sequence distance for inbound acked-delivery per authority.
   pub ack_receive_window:     u64,
 }
